Switch to zk-SNARK based on Groth's generic group construction
This post describes a vulnerability in the proving system used by Zcash,
which is implemented in a fork of libsnark's r1cs_ppzksnark:

https://z.cash/blog/zcash-counterfeiting-vulnerability-successfully-remediated/

The post indicates that the original libsnark implementation used by
Pequin is unaffected.

However, switching to the proving system due to Groth16 provides some
performance benefits.

Performance comparisons can be found here:

https://github.com/scipr-lab/libsnark/tree/master/libsnark/zk_proof_systems/ppzksnark